some cleans comments and moves cleanup volumes into helpers
